Fixes #17: restructure code
