+++ src/commands/replyTemplates.js
@@ -1,4 +1,12 @@
-/* eslint-disable import/prefer-default-export */
+export const contactInfo = `
+При обнаружении ошибок, пожалуйста, свяжитесь со мной.
+
+Contacts:
+---------
+Telegram: @mkangoo
+Mail: mkangoo24@gmail.com
+`
+
 export const greetingMessage = `
 Я - твой персональный бот,помогу быстро узнать расписание📆  
 Не теряй время на поиски, спроси меня! 🕵️‍♂️
