--- src/contactMe.js
@@ -1,8 +1,0 @@
-export const contactInfo = `
-При обнаружении ошибок, пожалуйста, свяжитесь со мной.
-
-Contacts:
----------
-Telegram: @mkangoo
-Mail: mkangoo24@gmail.com
-`
