+++ src/inlineKeyboard/inlineKeyboard.js
@@ -1,0 +1,5 @@
+import { Markup } from 'telegraf'
+
+export const getMainMenu = () => {
+	return Markup.keyboard([['Сегодня', 'Завтра'], ['Текущая неделя'], ['Следующая неделя']]).resize()
+}
